requirements.txt
flask==3.0.2
python-dotenv==0.19.0
requests==2.26.0